Change: visit系のメソッドをひとまとめに
