+++ spec/rails_helper.rb
@@ -63,4 +63,5 @@ RSpec.configure do |config|
 
   config.include FactoryBot::Syntax::Methods
   config.include LoginMacros
+  config.include VisitMacros
 end
